Interceptor initial mock up
In the future, each Interceptor will need to be an Entity so it
can be referenced by IEffectPacks.

This is a good start as it describes the base operation.

I'd like to get some tests in, switch to IEntity, and be able
to demonstrate transitivity in a test.
ie, I'd like a test where we register an interceptor to do a thing
then it is removed at the end of the turn by another interceptor
that was concurrently registered to remove it.