Effect checkpoint effort to decentralize definitions
Currently, all Effects need to be defined in the same location
modifying the same enum. This moves to a less central method.

We lose some type-safety but hopefully not as much as would appear
at first.
For example, we should be able to ensure that StartTurn.Self is
the only way to reference a StartTurn Effect, so we still have
that carrying us.