GameState remove Phase
Phase is an unnecessary concept given the introduction of
interceptors. The purpose of Phase was to be able to handle
lifecycles of 'non-immediate' Effects. Since interceptors can
handle EndTurn, StartTurn, and whatever other Phase 'markers' we
declare, there is no need for Phase to exist.

This touches IFilterState as that was a field that was being
filtered.