SetIntercept introduce optional Expiry semantics
This will allow SetIntercept to completely control its lifecycle
when registered. The dependency on knowing the identity of an
interceptor to target it with RemoveIntercept is removed by
only allow SetIntercept's expiry Mutator to Affix in a
RemoveIntercept Effect.

InterceptThrowGuardAndRemove relies on the Expiry semantics
established by this change. Previously, I thought there was an
easy solution to get around the identity and keep our declarativity
but tracing out the dependency has made it clear this is needed.