InterceptThrowGuardAndRemove fully implement
This implements the full lifecycle of an interceptor.
It is added, it does its job, and it gets removed on its expiry
condition.

PlayerPriority is used as the Expiry condition as it effectively
nops and we already have strong backing structure for making it
simple to use.

Machine is modified to remove null IEffectPacks early as mutators
may return null results which make the Filters unhappy. This is
desirable as we only want to allow explicit nulls.