InterceptForExtraTurn MultiEffect test checkpoint progress
This test ensures that we are able to tamper with a complicated
Effect, StartTurn, using an Intercept in order to get coherent
results.

Broken as many player-priority Events are getting added, which
is unexpected. The FilterState is throwing when it reaches the
fact that the currentTurn is mismatched when the real issue
is that the stack is never actually getting cleared out.