GameState initialize with Players Zone + Zone refactor LazyInit
The previous lazy initialization of Zones was causing issues as
we had a transitive, circular dependency resulting in the import
to Players in the Zones to be `undefined`.

GameState being initiailized with the Players Zone is a big step
as it will allow us to migrate the rest of the system to use it.