FilterState migrate zoneHas to zonePathHas
This is opens us up to having two forms of matching when we migrate
interceptors. This is the path-based, single property equality
based approach which is extremely useful for easy testing.
The second approach will be replacing interceptorsHas with
zoneContains for a diffed result of exactly the structure we
expect to receive.

The result of this is two styles of testing are made available.
This should allow faster iteration when path based solutions are
reasonable, such as matching on Player health or similar.
Additionally, the exhaustive matching of exact structure
matching provides extreme confidence in the result of potentially
destructive operations.