Zone migrate TargetType to per-zone + Global zone introduce
TargetTypes being per-zone rather than being in a single enum
means we gain significant modularity in how entities are resolved.

Global represents a pseudo-zone used only to expose its TargetType.
The necessity of the global target type is to be able to mutate
global state which Entities don't reside in.

The fact that Global is a zone means we don't need to consider
global state manipulation a special case. To ensure we can't,
it is impossible to manipulate global.