GameState introduce entities field + WithZone entity introduce
This introduces a central store for entities. We still need to
migrate individual zones to rely on this. Tests will also need to
be migrated.

WithZone allows us to enforce that a bare IEntity has the zone
field.

SetIntercept_test has been modified as it uses deep-diff against
a literal which had the IWithZone fields added.