core Filter factor out matchers into individual functions
This is the first required step before we can introduce Filter
registration.

The idea here is that the functions we've factored out are
considered the 'core' matchers which will always be included.
The user will be able to register new matchers by providing a
function which will work against an extension of the
IEffectPackFilter interface.