core FilterTest allow state and FilterMatcher in TestCase
GameState can be provided to allow the testing of filters which
may conditionally match based on current GameState.
This needs to be used carefully. Prefer purely Effect-based
filters when possible. Viewing the GameState is powerful but can
allow the development of very difficult to maintain Filters.

ie, This allows a 'PassesGo' filter in Monopoly which relies on
being able to query the position of a player. This should always
be gated around the appropriate Effect-based context; ie,
requiring an Effect of Move.

FilterMatcher's inclusion allows this to be applicable to non-core
casese. This follows the pattern established by Effect and Mutator