monoply PassGo filter introduce + FilterTest structure declare
This PassGo filter, along with the property-copying Affix mutator,
will allow us to factor the 'passing go' functionality currently
kept in the Move effect into an interceptor.
Hence, the game rules will be more declarative.

This is the first instance of testing a non-core Filter, I think
the structure feels pretty good so far.