core MultiEffectTest more helpful handling of unexpected errors
Previously, the error would bubble to the top-level test try
which would print the error without the case name or stack trace.

This change ensures that we get all the context that was previously
lost. This is especially useful for when an effect which has
not been registered has been used, which throws.